release: Chromium 46.0.2490 (46.0.2490.92)
Compare: https://github.com/chromium/chromium/compare/94ba9a023a52ad3b52bd144d4070168c8181adb5...08ceb3910e2af5bff59cba8ba7fe118830a83592

+++ chrome/android/java/src/org/chromium/chrome/browser/document/DocumentMigrationHelper.java
@@ -28,7 +28,6 @@ import org.chromium.chrome.R;
 import org.chromium.chrome.browser.ApplicationLifetime;
 import org.chromium.chrome.browser.ChromeApplication;
 import org.chromium.chrome.browser.IntentHandler;
-import org.chromium.chrome.browser.Tab;
 import org.chromium.chrome.browser.TabState;
 import org.chromium.chrome.browser.UrlConstants;
 import org.chromium.chrome.browser.UrlUtilities;
@@ -41,11 +40,13 @@ import org.chromium.chrome.browser.favicon.FaviconHelper.FaviconImageCallback;
 import org.chromium.chrome.browser.ntp.NativePageFactory;
 import org.chromium.chrome.browser.preferences.ChromePreferenceManager;
 import org.chromium.chrome.browser.profiles.Profile;
+import org.chromium.chrome.browser.tab.Tab;
 import org.chromium.chrome.browser.tab.TabIdManager;
 import org.chromium.chrome.browser.tabmodel.TabCreatorManager;
 import org.chromium.chrome.browser.tabmodel.TabPersistentStore;
 import org.chromium.chrome.browser.tabmodel.TabPersistentStore.OnTabStateReadCallback;
 import org.chromium.chrome.browser.tabmodel.document.ActivityDelegate;
+import org.chromium.chrome.browser.tabmodel.document.ActivityDelegateImpl;
 import org.chromium.chrome.browser.tabmodel.document.DocumentTabModel;
 import org.chromium.chrome.browser.tabmodel.document.DocumentTabModel.Entry;
 import org.chromium.chrome.browser.tabmodel.document.DocumentTabModel.InitializationObserver;
@@ -101,7 +102,7 @@ public class DocumentMigrationHelper {
      * Stores a list of "tasks" that are meant to be returned by the ActivityManager for migration.
      * The tasks are inserted manually during the migration from classic mode to document mode.
      */
-    private static class MigrationActivityDelegate extends ActivityDelegate {
+    private static class MigrationActivityDelegate extends ActivityDelegateImpl {
         private final List<Entry> mEntries;
         private final int mSelectedTabId;
 
@@ -203,12 +204,6 @@ public class DocumentMigrationHelper {
             mFinalizeMode = finalizeMode;
         }
 
-        protected void finalize() throws Throwable {
-            // TODO(dfalcantara): Remove this log.  crbug.com/513130
-            Log.w(TAG, "Finalizing MigrationImageCallback: " + this);
-            super.finalize();
-        }
-
         @Override
         public void onFaviconAvailable(final Bitmap favicon, String iconUrl) {
             mFavicon = favicon;
@@ -217,10 +212,6 @@ public class DocumentMigrationHelper {
 
         @Override
         public void onFinishGetBitmap(Bitmap bitmap) {
-            // TODO(dfalcantara): Remove this log.  crbug.com/513130
-            Log.w(TAG, "MigrationImageCallback.onFinishGetBitmap: " + this + ", "
-                    + mCompletedCount.get() + "/" + mTabModel.getCount());
-
             // Even if the favicon or thumbnail are null, we still add the AppTask: the framework
             // handles null favicons and addAppTask() below handles null thumbnails.
             if (!NativePageFactory.isNativePageUrl(mUrl, false)
@@ -277,21 +268,23 @@ public class DocumentMigrationHelper {
         // All the TabStates (incognito or not) live in the same directory.
         File[] allTabs = normalTabModel.getStorageDelegate().getStateDirectory().listFiles();
         try {
-            for (int i = 0; i < allTabs.length; i++) {
-                String fileName = allTabs[i].getName();
-                Pair<Integer, Boolean> tabInfo = TabState.parseInfoFromFilename(fileName);
-                if (tabInfo == null) continue;
-                int tabId = tabInfo.first;
+            if (allTabs != null) {
+                for (int i = 0; i < allTabs.length; i++) {
+                    String fileName = allTabs[i].getName();
+                    Pair<Integer, Boolean> tabInfo = TabState.parseInfoFromFilename(fileName);
+                    if (tabInfo == null) continue;
+                    int tabId = tabInfo.first;
 
-                // Also remove the tab state file for the closed tabs.
-                boolean success;
-                if (!tabIdsToRemove.contains(tabId)) {
-                    success = allTabs[i].renameTo(new File(migratedFolder, fileName));
-                } else {
-                    success = allTabs[i].delete();
+                    // Also remove the tab state file for the closed tabs.
+                    boolean success;
+                    if (!tabIdsToRemove.contains(tabId)) {
+                        success = allTabs[i].renameTo(new File(migratedFolder, fileName));
+                    } else {
+                        success = allTabs[i].delete();
+                    }
+
+                    if (!success) Log.e(TAG, "Failed to move/delete file for tab ID: " + tabId);
                 }
-
-                if (!success) Log.e(TAG, "Failed to move/delete file for tab ID: " + tabId);
             }
 
             if (normalTabModel.getCount() != 0) {
@@ -349,7 +342,9 @@ public class DocumentMigrationHelper {
      */
     public static boolean migrateTabsToDocumentForUpgrade(Activity activity,
             int finalizeMode) {
-        // Temporarily allowing disk access. TODO: Fix. See http://crbug.com/493157
+        // Allow StrictMode violations here as state file read/write is on the critical path. This
+        // is also a one-time migration, so optimizing further does not improve daily experience.
+        // See http://crbug.com/493157 for more context.
         StrictMode.ThreadPolicy oldPolicy = StrictMode.allowThreadDiskReads();
         StrictMode.allowThreadDiskWrites();
         try {
@@ -396,10 +391,11 @@ public class DocumentMigrationHelper {
         // Create maps for all tabs that will be used during TabModel initialization.
         final List<Entry> normalEntryMap = new ArrayList<Entry>();
 
-        int currentSelectorIndex = 0;
-        File currentFolder = TabPersistentStore.getStateDirectory(activity, currentSelectorIndex);
+        File currentFolder = null;
         MigrationTabStateReadCallback callback = new MigrationTabStateReadCallback();
-        while (currentFolder.listFiles() != null && currentFolder.listFiles().length != 0) {
+        for (int currentSelectorIndex = 0;
+                (currentFolder = getNextNonEmptyFolder(activity, currentSelectorIndex)) != null;
+                ++currentSelectorIndex) {
             File[] allTabs = TabPersistentStore
                     .getStateDirectory(activity, currentSelectorIndex).listFiles();
             try {
@@ -408,27 +404,33 @@ public class DocumentMigrationHelper {
                 Log.e(TAG, "IO Exception while trying to get the last used tab id");
             }
 
-            for (int i = 0; i < allTabs.length; i++) {
-                // Move tab state file to the document side folder.
-                String fileName = allTabs[i].getName();
-                Pair<Integer, Boolean> tabInfo = TabState.parseInfoFromFilename(fileName);
-                if (tabInfo == null) continue;
+            if (allTabs != null) {
+                for (int i = 0; i < allTabs.length; i++) {
+                    // Move tab state file to the document side folder.
+                    String fileName = allTabs[i].getName();
+                    Pair<Integer, Boolean> tabInfo = TabState.parseInfoFromFilename(fileName);
+                    if (tabInfo == null) continue;
 
-                boolean success;
-                if (tabInfo.second) {
-                    success = allTabs[i].delete();
-                } else {
-                    success = allTabs[i].renameTo(new File(migratedFolder, fileName));
-                    normalEntryMap.add(new Entry(tabInfo.first, UrlConstants.NTP_URL));
+                    boolean success;
+                    if (tabInfo.second) {
+                        success = allTabs[i].delete();
+                    } else {
+                        success = allTabs[i].renameTo(new File(migratedFolder, fileName));
+                        normalEntryMap.add(new Entry(tabInfo.first, UrlConstants.NTP_URL));
+                    }
+
+                    if (!success) Log.e(TAG, "Failed to move/delete file: " + fileName);
                 }
-
-                if (!success) Log.e(TAG, "Failed to move/delete file: " + fileName);
             }
-            currentSelectorIndex++;
-            currentFolder = TabPersistentStore.getStateDirectory(activity, currentSelectorIndex);
         }
 
         return new MigrationActivityDelegate(normalEntryMap, callback.getSelectedTabId());
+    }
+
+    private static File getNextNonEmptyFolder(Activity activity, int currentSelectorIndex) {
+        File folder = TabPersistentStore.getStateDirectory(activity, currentSelectorIndex);
+        File[] files = folder.listFiles();
+        return (files == null || files.length == 0) ? null : folder;
     }
 
     private static void addAppTasksFromFiles(final Activity activity,
@@ -515,7 +517,7 @@ public class DocumentMigrationHelper {
      */
     public static void migrateTabs(boolean toDocumentMode, final Activity activity,
             boolean terminate) {
-        // Temporarily allowing disk access. TODO: Fix. See http://crbug.com/493157
+        // Allowing StrictMode violations as above. See http://crbug.com/493157 for context.
         StrictMode.ThreadPolicy oldPolicy = StrictMode.allowThreadDiskReads();
         StrictMode.allowThreadDiskWrites();
         try {
